refactor: move blogs index to index
